Rename metadata-browser -> metadata_browser I think this is the accepted convention?
