+++ src/metadata_browser/MetadataComponent.ts
